Editor: split BigPropertySheet parent class of GeneralSettings pane
